codes/requirements.txt
absl-py==0.9.0
aiohttp 
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.15.0
chardet 
charset-normalizer 
click==8.0.3
cryptography
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
idna-ssl 
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.3.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict 
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect 
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.12.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT 
pyOpenSSL 
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks
pyte==0.8.0
python-dateutil==2.8.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.8.0
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thefuck==3.32
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
yarl 
zipp==3.1.0
